requirements.txt
torch
torchvision
albumentation
scikit-learn
tqdm
